requirements.txt
# Core dependencies
fastapi[all]==0.115.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Google Drive integration
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=9.0.0
pytesseract>=0.3.8

# LlamaIndex and Azure Search
llama-index>=0.10.0
llama-index-vector-stores-azureaisearch>=0.1.0
openai>=1.12.0

# Monitoring and logging
prometheus-client>=0.16.0
python-json-logger>=2.0.7

# Additional dependencies
uvicorn>=0.15.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
alembic>=1.7.0 
